% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCStSNE.R
\name{FCStSNE}
\alias{FCStSNE}
\title{tSNE and OneSENSE algorithm for FCS data}
\usage{
FCStSNE(LoaderPATH = "fcs", ceil = 5000, FNnames = "names.csv",
  OutputSuffix = "Out", DotSNE = TRUE, DoOneSENSE = TRUE, Bins = 250)
}
\arguments{
\item{LoaderPATH}{Path where FCS file is located}

\item{ceil}{Maximum number of cells to sample from each fcs sample/file}

\item{FNnames}{.csv file generated when markers from each category are selected}

\item{OutputSuffix}{suffix to name output folder}

\item{DotSNE}{boolean, if TRUE do tSNE, if FALSE skip tSNE}

\item{DoOneSENSE}{boolean, if TRUE do OneSENSE, if FALSE skip OneSENSE}

\item{Bins}{number of bins to put the cell data into, DEFAULT = 250}
}
\value{
FCS files, tSNE histograms, OneSENSE plot
}
\description{
tSNE and OneSENSE algorithm for FCS data
}
\examples{
#dir <- system.file('extdata',package='oneSENSE')
#fnnames <- system.file('extdata', "names.csv", package='oneSENSE')

#FCStSNE(LoaderPATH = dir, FNnames = fnnames) #remove hash symbol to run
}
