% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YKOneSmapperFlour.R
\name{getCoords}
\alias{getCoords}
\title{Get Coordinates of median position}
\usage{
getCoords(LoaderPATH = LoaderPATH, FFdata = FFdata)
}
\arguments{
\item{LoaderPATH}{Path of fcs_out files}

\item{FFdata}{Flow frame data}
}
\value{
Assign global variables for access
}
\description{
Get Coordinates of median position
}
\examples{
#remove hash symbol to run
dir4 <- system.file('extdata/extra', package = 'oneSENSE')
file5 <- system.file('extdata/myFFdatas.Rds', package = 'oneSENSE')
FFdata = readRDS(file5)
getCoords(dir4, FFdata)
}
