% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YKOneSmapperFlour.R
\name{getParameters}
\alias{getParameters}
\title{Get parameters from FCS files}
\usage{
getParameters(rawFCSdir)
}
\arguments{
\item{rawFCSdir}{FCS path}
}
\value{
mid
}
\description{
Get parameters from FCS files
}
\examples{
#remove hash to run
dir3 <- system.file('extdata/fcs', package = 'oneSENSE')
getParameters(dir3)
}
