% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneSENSE_GUI.R
\name{getParameters_GUI}
\alias{getParameters_GUI}
\title{GUI for marker selection}
\usage{
getParameters_GUI(fcsFile, rawFCSdir)
}
\arguments{
\item{fcsFile}{The name of the FCS file}

\item{rawFCSdir}{The path of the FCS file}
}
\value{
a dataframe containing selected markers per catergory
}
\description{
Extract the markers from the fcsfiles
}
\examples{
#getParameters_GUI()
}
