% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneSENSE_GUI.R
\name{onesense_GUI}
\alias{onesense_GUI}
\title{A user friendly GUI client for oneSENSE package}
\usage{
onesense_GUI()
}
\value{
the GUI for oneSENSE analysis
}
\description{
This GUI provides an easy way for Flow Cytometry data analysis using the oneSENSE package. Main parameters for running 'oneSENSE' were integrated in this GUI, and analysis results are launched in Rstudio after submission.
}
\examples{
onesense_GUI()
}
\author{
Yong kee Tan
}
