% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YKOneSmapperFlour.R
\name{OneSmapperFreq1}
\alias{OneSmapperFreq1}
\title{Generate Frequency Heatplot Part 1}
\usage{
OneSmapperFreq1(LoaderPATH = "fcs_Out")
}
\arguments{
\item{LoaderPATH}{Path of fcs_Out files}
}
\value{
Flow Frame data for coordinate selection
}
\description{
Generate Frequency Heatplot Part 1
}
\examples{
dir1 <- system.file('extdata/fcs',package='oneSENSE')
OneSmapperFreq1(dir1) #remove hash symbol to run
}
