% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YKOneSmapperFlour.R
\name{getParameters}
\alias{getParameters}
\title{Title
Get parameters}
\usage{
getParameters(rawFCSdir)
}
\arguments{
\item{rawFCSdir}{FCS path}
}
\value{
mid
}
\description{
Title
Get parameters
}
\examples{
#remove hash to run
dir3 <- system.file('extdata', package = 'oneSENSE')
getParameters(dir3)
}
