% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneSENSE_GUI.R
\name{oneSENSE_GUI}
\alias{oneSENSE_GUI}
\title{A user friendly GUI client for oneSENSE package}
\usage{
oneSENSE_GUI()
}
\value{
GUI for onesense analysis
}
\description{
This GUI provides an easy way for Flow Cytometry data analysis
using the oneSENSE package. Main parameters for running 'oneSENSE'
were integrated in this GUI, and analysis results are launched
in Rstudio after submission.
}
\examples{

if (interactive()) oneSENSE::oneSENSE_GUI()
}
