% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YKOneSmapperFlour.R
\name{OneSmapperFreq1}
\alias{OneSmapperFreq1}
\title{Title}
\usage{
OneSmapperFreq1(LoaderPATH = "fcs")
}
\arguments{
\item{LoaderPATH}{Path of fcs_Out files}
}
\value{
Flow Frame data for coordinate selection
}
\description{
Title
}
\examples{
dir1 <- system.file('extdata',package='oneSENSE')
OneSmapperFreq1(dir1) #remove hash symbol to run
}
