% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YKOneSmapperFlour.R
\name{OneSmapperFreq2}
\alias{OneSmapperFreq2}
\title{Title
Frequency Heatplot Generation Part 2}
\usage{
OneSmapperFreq2(LoaderPATH = "fcs", Bins = 250)
}
\arguments{
\item{LoaderPATH}{path of fcs_Out files}

\item{Bins}{Number of bins}
}
\value{
returns frequency heatplot
}
\description{
Title
Frequency Heatplot Generation Part 2
}
\examples{
dir2 <- system.file('extdata',package='oneSENSE')

#OneSmapperFreq2(dir2, 250) #remove hash symbol to run
}
